requirements.txt
asgiref==3.7.2
astroid==3.0.3
blinker==1.7.0
certifi==2023.11.17
cfgv==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.8
distlib==0.3.8
Django==5.0.1
filelock==3.13.1
Flask==3.0.2
identify==2.5.33
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
nodeenv==1.8.0
pillow==10.2.0
pipenv==2023.11.15
platformdirs==4.1.0
pre-commit==3.6.0
pylint==3.0.3
python-barcode==0.15.1
PyYAML==6.0.1
setuptools==69.0.3
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.4
virtualenv==20.25.0
Werkzeug==3.0.1
